+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   and `:code`. An update with a `Sourceror.Zipper.zipper()` is no longer
   supported.
 + Add `Rewrite.Source.from_ast/3`.
++ Add `Rewrite.Source.owner/1`.
 
 ## 0.1.1 - 2022/09/07